Use new addMetadata method.
`zim::Creator` now has a specific method to add a metadata, let's use it.
