+++ src/zimwriterfs/article.cpp
@@ -149,41 +149,3 @@ RedirectArticle::RedirectArticle(const ZimCreatorFS *_creator,
   this->redirectUrl = redirectUrl;
   mimeType = getMimeTypeForFile(creator->basedir(), redirectUrl.getUrl());
 }
-
-SimpleMetadataArticle::SimpleMetadataArticle(const std::string& id,
-                                             const std::string& value)
-    : MetadataArticle(id), value(value)
-{
-}
-
-MetadataDateArticle::MetadataDateArticle() : MetadataArticle("Date")
-{
-}
-
-void MetadataDateArticle::genDate() const
-{
-  time_t t = time(0);
-  struct tm* now = localtime(&t);
-  std::stringstream stream;
-  stream << (now->tm_year + 1900) << '-' << std::setw(2) << std::setfill('0')
-         << (now->tm_mon + 1) << '-' << std::setw(2) << std::setfill('0')
-         << now->tm_mday;
-  data = stream.str();
-}
-
-zim::Blob MetadataDateArticle::getData() const
-{
-  if (data.empty()) {
-    genDate();
-  }
-  return zim::Blob(data.data(), data.size());
-}
-
-zim::size_type MetadataDateArticle::getSize() const
-{
-  if(data.empty()) {
-    genDate();
-  }
-  return data.size();
-}
-
