+++ src/zimwriterfs/article.h
@@ -47,76 +47,6 @@ class Article : public zim::writer::Article
   virtual ~Article(){};
 };
 
-class MetadataArticle : public zim::writer::Article
-{
- protected:
-   std::string id;
-
- public:
-  explicit MetadataArticle(const std::string& id) : id(id) {};
-  virtual zim::writer::Url getUrl() const { return zim::writer::Url('M', id); }
-  virtual zim::writer::Url getRedirectUrl() const { return zim::writer::Url(); }
-  virtual bool isInvalid() const { return false; }
-  virtual std::string getTitle() const { return ""; }
-  virtual bool isRedirect() const { return false; }
-  virtual bool isLinktarget() const { return false; }
-  virtual bool isDeleted() const { return false; }
-  virtual std::string getMimeType() const { return "text/plain"; }
-  virtual bool shouldIndex() const { return false; }
-  virtual bool shouldCompress() const { return true; }
-  virtual std::string getFilename() const { return ""; }
-};
-
-class SimpleMetadataArticle : public MetadataArticle
-{
- private:
-  std::string value;
-
- public:
-  explicit SimpleMetadataArticle(const std::string& id,
-                                 const std::string& value);
-  virtual zim::Blob getData() const
-  {
-    return zim::Blob(value.c_str(), value.size());
-  }
-  virtual zim::size_type getSize() const
-  {
-    return value.size();
-  }
-};
-
-/// This class creates a redirect entry to image/png favicon
-class MetadataFaviconArticle : public MetadataArticle
-{
- private:
-   zim::writer::Url redirectUrl;
- public:
-  explicit MetadataFaviconArticle(zim::writer::Url value)
-    : MetadataArticle(""), redirectUrl(value) {}
-  virtual zim::writer::Url getUrl() const { return zim::writer::Url('-', "favicon"); }
-  virtual bool isInvalid() const { return false; }
-  virtual std::string getTitle() const { return ""; }
-  virtual bool isRedirect() const { return true; }
-  virtual std::string getMimeType() const { return "image/png"; }
-  virtual zim::writer::Url getRedirectUrl() const { return redirectUrl; }
-  virtual bool shouldCompress() const { return false; }
-  virtual std::string getFilename() const { return ""; }
-  virtual zim::Blob getData() const { return zim::Blob(); }
-  virtual zim::size_type getSize() const { return 0; }
-};
-
-class MetadataDateArticle : public MetadataArticle
-{
- private:
-  mutable std::string data;
-  void genDate() const;
-
- public:
-  MetadataDateArticle();
-  virtual zim::Blob getData() const;
-  virtual zim::size_type getSize() const;
-};
-
 class FileArticle : public Article
 {
  private:
