+++ src/zimwriterfs/tools.h
@@ -28,5 +28,6 @@
 std::string extractRedirectUrlFromHtml(const GumboVector* head_children);
 void getLinks(GumboNode* node, std::map<std::string, bool>& links);
 
+std::string generateDate();
 
 #endif  // OPENZIM_ZIMWRITERFS_TOOLS_H
